LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-focus-expected.txt


FAIL :focus via :scope in subject assert_equals: expected "1" but got "auto"
PASS :focus via :scope in non-subject
FAIL :focus in limit, :scope in subject assert_equals: expected "auto" but got "1"
PASS :focus in intermediate limit, :scope in subject
